refactor(tests): removing "mock" from file names
