+++ test/mocks/data/firebase_auth_user.dart
