+++ test/mocks/overrides/override_auth_providers.dart
@@ -5,7 +5,7 @@ import "package:google_sign_in_mocks/google_sign_in_mocks.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/services/login_service.dart";
 
-import "../data/mock_firebase_auth_user.dart";
+import "../data/firebase_auth_user.dart";
 
 /// Override to use to test the login process
 final firebaseAuthMocksOverrides = [
