+++ test/models/database/user/user_data_test.dart
@@ -1,7 +1,7 @@
 import "package:flutter_test/flutter_test.dart";
 import "package:proxima/models/database/user/user_data.dart";
 
-import "../../../mocks/data/mock_firestore_user.dart";
+import "../../../mocks/data/firestore_user.dart";
 
 void main() {
   group("User Data testing", () {
