+++ test/services/database/user_repository_test.dart
@@ -6,7 +6,7 @@ import "package:proxima/models/database/user/user_firestore.dart";
 import "package:proxima/models/database/user/user_id_firestore.dart";
 import "package:proxima/services/database/user_repository_service.dart";
 
-import "../../mocks/data/mock_firestore_user.dart";
+import "../../mocks/data/firestore_user.dart";
 
 void main() {
   group("User repository testing", () {
